requirements.txt
discord.py
pymongo[srv]
pandas